requirements.txt
colorama==0.4.6
faiss_cpu==1.10.0
jieba==0.42.1
numpy==2.2.4
openai==1.68.2
Pillow==11.1.0
pymongo==4.11.3
Requests==2.32.3
websockets==15.0.1
